feat(master): 接入 card2 verify 结果联动跟随
WHAT:
- 让主板采集 card2 verify 快照并把识别状态纳入 tracking input 输出
- 在 TRACK_START 前主动向 card2 请求重置 face session
- 调整 card1/card2 日志和 selected_idx 表达，兼容新的跟随状态语义

WHY:
- 仅有 card2 子板侧识别摘要还不够，主板必须消费 verify 状态才能驱动后续跟随丢失、等待验证和找回逻辑
- 每次跟随开始前如果不重置 face session，旧模板可能污染新一轮目标确认
- tracking input 需要同时表示 card1 跟随状态和 card2 verify 结果，才能为上层控制提供稳定语义

HOW:
- 在 app_card2_result_handler 和 app_card2_subboard 中新增 verify snapshot 读取与 FACE_SESSION_RESET 请求接口
- 在 app_gimbal_tracking_input 中扩展 follow state、verify 字段、follow_session_id 和 using_card2_verify 语义，并合并 card2 结果到输出
- 在 app_action_executor 中把 TRACK_START 接到 face session reset，同时修正主板结果日志与 selected_idx 类型

+++ Projects/App_GimbalMaster/Src/app_card1_result_handler.c
@@ -89,11 +89,11 @@ void app_card1_result_handler_tick(void)
     if (!s_overlay_active && overlay_now_active) {
         if ((s_last_overlay_log_ms == 0u) ||
             ((uint32_t)(now_ms - s_last_overlay_log_ms) >= CARD1_OVERLAY_LOG_MIN_INTERVAL_MS)) {
-            MASTER_LOG_INFO("[MASTER][CARD1][OVL] source=%s frame=%lu target=%u selected=%u effective=%s raw_tracker=%s raw_flags=0x%02X id=%u conf=%u miss=%u pred=%u box=(%ld,%ld)-(%ld,%ld)\r\n",
+            MASTER_LOG_INFO("[MASTER][CARD1][OVL] source=%s frame=%lu target=%u selected=%ld effective=%s raw_tracker=%s raw_flags=0x%02X id=%u conf=%u miss=%u pred=%u box=(%ld,%ld)-(%ld,%ld)\r\n",
                             output.from_tracking_summary ? "summary" : "result-fallback",
                             (unsigned long)output.frame_id,
                             output.valid ? 1u : 0u,
-                            (unsigned)output.selected_idx,
+                            (long)output.selected_idx,
                             app_gimbal_tracking_input_view_name(output.effective_view),
                             output.raw_state_valid ? app_gimbal_tracking_input_raw_tracker_name(output.tracker_state_raw) : "UNKNOWN",
                             (unsigned)output.tracker_flags_raw,
@@ -110,11 +110,11 @@ void app_card1_result_handler_tick(void)
     } else if (s_overlay_active && !overlay_now_active) {
         if ((s_last_overlay_log_ms == 0u) ||
             ((uint32_t)(now_ms - s_last_overlay_log_ms) >= CARD1_OVERLAY_LOG_MIN_INTERVAL_MS)) {
-            MASTER_LOG_INFO("[MASTER][CARD1][OVL] cleared source=%s frame=%lu target=%u selected=%u effective=%s raw_tracker=%s raw_flags=0x%02X id=%u conf=%u miss=%u pred=%u\r\n",
+            MASTER_LOG_INFO("[MASTER][CARD1][OVL] cleared source=%s frame=%lu target=%u selected=%ld effective=%s raw_tracker=%s raw_flags=0x%02X id=%u conf=%u miss=%u pred=%u\r\n",
                             output.from_tracking_summary ? "summary" : "result-fallback",
                             (unsigned long)output.frame_id,
                             output.valid ? 1u : 0u,
-                            (unsigned)output.selected_idx,
+                            (long)output.selected_idx,
                             app_gimbal_tracking_input_view_name(output.effective_view),
                             output.raw_state_valid ? app_gimbal_tracking_input_raw_tracker_name(output.tracker_state_raw) : "UNKNOWN",
                             (unsigned)output.tracker_flags_raw,
